remove internal generic types
