+++ page_test.go
@@ -1,5 +1,3 @@
-//go:build !go1.18
-
 package parquet_test
 
 import (
